Fix link logo rules being too broad.
Reddit does stupid link rewriting. During the rewriting stage, the
original url is in the data-href-url attribute. This commit adds rules
to match this, and tightens up the targets.

Also removes the rule to match uppercase subreddit links, in an attempt
to reclaim a tiny bit of lost space. I don't think anyone needs to link
/r/SDFLEET. This does lose /r/AAFB but that can be rectified by using
the upper case version as the main url for AAF (since we force the lower
case version).